pipeline_demo/i2j_ui/requirements.txt
# Web UI backend
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Reuse OCR/LLM pipeline deps
requests>=2.28.0
Pillow>=10.0.0
jsonschema>=4.0.0
numpy>=1.20.0

# OCR engines (match pipeline)
easyocr>=1.7.0
pytesseract>=0.3.10
opencv-python-headless>=4.5.0

# Google Sheets (optional)
gspread>=6.0.0
google-auth>=2.35.0